requirements.txt
matplotlib==3.3.2
pandas==1.1.3
Pillow==9.4.0
plotly==5.13.0
seaborn==0.11.0
streamlit==1.17.0
